Metabolomics MSnSet

An object of class MSnbase::MSnSet containing log\(_2\)-transformed metabolite data for differential analysis and WGCNA.

Usage

METAB_MSNSET

Format

Object of class MSnbase::MSnSet with 1063 features and 50 samples.

assayData

Object of class matrix with 1063 rows and 50 columns. See METAB_NORM_DATA_NESTED for more details.

featureData

Object of class data.frame with 1063 rows and 13 variables containing additional feature information:

feature_ID

character; metabolite RefMet identifier.

dataset

character; metabolomics platform in which the feature was measured.

name_in_figures

character; alternative feature identifier primarily used in visualizations.

refmet_super_class

character; RefMet chemical super class.

refmet_main_class

character; RefMet chemical main class.

refmet_sub_class

character; RefMet chemical sub class.

lipid_class

character; (lipids only) the class of the lipid. There are 43 unique classes.

chain_length

integer; (lipids only) total number of carbons that comprise the fatty acid chains.

double_bond

integer; total number of C=C double bonds in the fatty acid chains.

rt

numeric; retention time. Used to assign lipid_class.

mz

numeric; mass-to-charge ratio (m/z).

neutral_mass

numeric; neutral mass (g/mol).

formula

character; chemical formula.

phenoData

Object of class data.frame with 50 rows and 6 variables containing additional sample information:

pid

integer; randomly generated 8-digit identifier used in linkage to phenotypic data. All samples from the same animal have the same PID.

bid

integer; unique 5 digit identifier of all samples collected for an acute test/sample collection period. All samples collected during that period will have the same BID.

labelid

integer; unique 11 digit specimen label identifier, originating at the collection site, that provides a link to specimen processing and is used for shipments to the biorepository. Same as viallabel only in instances where aliquots are not further processed at the biorepository.

sex

factor; the sex of the rat with levels "Female" and "Male".

timepoint

factor; exercise training group. Either "SED" (sedentary) or the number of weeks of training ("1W", "2W", "4W", "8W").

exp_group

factor; experimental group. Unique combination of sex (first letter) and timepoint.

References

Fahy, E., & Subramaniam, S. (2020). RefMet: A reference nomenclature for metabolomics. Nature Methods, 17(12), 1173--1174. https://doi.org/10.1038/s41592-020-01009-y

See also

Examples

library(MSnbase)
#> Error in library(MSnbase): there is no package called ‘MSnbase’
METAB_MSNSET                   # summary
#> Loading required package: MSnbase
#> Warning: there is no package called ‘MSnbase’
#> Error in .requirePackage(package): unable to find required package ‘MSnbase’
exprs(METAB_MSNSET)[1:10, 1:5] # assayData (first 10 rows and 5 columns)
#> Error in exprs(METAB_MSNSET): could not find function "exprs"
head(fData(METAB_MSNSET))      # featureData
#> Error in fData(METAB_MSNSET): could not find function "fData"
head(pData(METAB_MSNSET))      # phenoData
#> Error in pData(METAB_MSNSET): could not find function "pData"