Phosphoproteomics MSnSet

An object of class MSnbase::MSnSet containing median-MAD-normalized log\(_2\)-transformed protein phosphorylation data for differential analysis.

Usage

PHOSPHO_MSNSET

Format

Object of class MSnbase::MSnSet with 30304 features and 60 samples.

assayData

Object of class matrix with 30304 rows and 60 columns. See PHOSPHO_WATSC_NORM_DATA for more details.

featureData

Object of class data.frame with 30304 rows and 7 variables containing additional feature information:

feature_ID

character; Reference Sequence (RefSeq) protein identifier followed by an underscore and position(s) of phosphorylation.

gene_symbol

character; official gene symbol.

entrez_gene

integer; Entrez gene identifier.

site

character; phosphorylation site(s) extracted from feature_ID column and stripped of trailing lowercase amino acid letters (s, t, y). Multiple sites are separated by semicolons.

human_feature_ID

character; human UniProt accession followed by an underscore and positions of phosphorylation.

human_uniprot

character; human UniProt accession extracted from human_feature_ID column.

human_site

character; human phosphorylation site(s) extracted from human_feature_ID column and stripped of trailing lowercase amino acid letters (s, t, y). Multiple sites are separated by semicolons.

phenoData

Object of class data.frame with 60 rows and 7 variables containing additional sample information:

pid

integer; randomly generated 8-digit identifier used in linkage to phenotypic data. All samples from the same animal have the same PID.

bid

integer; unique 5 digit identifier of all samples collected for an acute test/sample collection period. All samples collected during that period will have the same BID.

labelid

integer; unique 11 digit specimen label identifier, originating at the collection site, that provides a link to specimen processing and is used for shipments to the biorepository. Same as viallabel only in instances where aliquots are not further processed at the biorepository.

viallabel

character; unique 11 digit sample vial identifier. Starts with the bid.

sex

factor; the sex of the rat with levels "Female" and "Male".

timepoint

factor; exercise training group. Either "SED" (sedentary) or the number of weeks of training ("1W", "2W", "4W", "8W").

exp_group

factor; experimental group. Unique combination of sex (first letter) and timepoint.

Details

RAT_TO_HUMAN_PHOSPHO for details of mapping between rat and human phosphorylation sites.

FEATURE_TO_GENE for details of feature-to-gene mapping.

See also

Examples

library(MSnbase)
PHOSPHO_MSNSET                   # summary
#> MSnSet (storageMode: lockedEnvironment)
#> assayData: 30304 features, 60 samples 
#>   element names: exprs 
#> protocolData: none
#> phenoData
#>   sampleNames: 90222017006 90258017006 ... 90237017006 (60 total)
#>   varLabels: pid bid ... exp_group (7 total)
#>   varMetadata: labelDescription
#> featureData
#>   featureNames: NP_001001504.1_S448s NP_001001512.2_S228s ...
#>     XP_345868.4_T633t (30304 total)
#>   fvarLabels: feature_ID gene_symbol ... human_site (7 total)
#>   fvarMetadata: labelDescription
#> experimentData: use 'experimentData(object)'
#> Annotation:  
#> - - - Processing information - - -
#>  MSnbase version: 2.22.0 
exprs(PHOSPHO_MSNSET)[1:10, 1:5] # assayData (first 10 rows and 5 columns)
#>                      90222017006 90258017006 90441017006 90266017006
#> NP_001001504.1_S448s          NA          NA          NA          NA
#> NP_001001512.2_S228s      0.1112     -0.0098      0.1768     -0.2122
#> NP_001001801.1_S30s           NA          NA          NA          NA
#> NP_001002016.2_S12s      -0.0295      0.0111     -0.1142     -0.1185
#> NP_001002016.2_S143s          NA          NA          NA          NA
#> NP_001002016.2_S153s     -0.5996      0.3427     -0.3792      0.1080
#> NP_001002016.2_S17s      -0.2917      0.2004     -0.7424      0.1419
#> NP_001002016.2_S18s      -0.0802      0.1502     -0.1014      0.0237
#> NP_001002016.2_S212s          NA          NA          NA          NA
#> NP_001002016.2_S22s       0.1182     -0.1563      0.4189     -0.1975
#>                      90421017006
#> NP_001001504.1_S448s          NA
#> NP_001001512.2_S228s      0.2223
#> NP_001001801.1_S30s           NA
#> NP_001002016.2_S12s       0.0288
#> NP_001002016.2_S143s          NA
#> NP_001002016.2_S153s     -0.4339
#> NP_001002016.2_S17s      -0.4537
#> NP_001002016.2_S18s       0.0077
#> NP_001002016.2_S212s          NA
#> NP_001002016.2_S22s       0.1365
head(fData(PHOSPHO_MSNSET))      # featureData
#>                                feature_ID gene_symbol entrez_gene site
#> NP_001001504.1_S448s NP_001001504.1_S448s    Gtf2ird1      246770 S448
#> NP_001001512.2_S228s NP_001001512.2_S228s       Gtf2i      353256 S228
#> NP_001001801.1_S30s   NP_001001801.1_S30s       Akap7      361458  S30
#> NP_001002016.2_S12s   NP_001002016.2_S12s        Lmna       60374  S12
#> NP_001002016.2_S143s NP_001002016.2_S143s        Lmna       60374 S143
#> NP_001002016.2_S153s NP_001002016.2_S153s        Lmna       60374 S153
#>                      human_feature_ID human_uniprot human_site
#> NP_001001504.1_S448s     Q9UHL9_S448s        Q9UHL9       S448
#> NP_001001512.2_S228s             <NA>          <NA>       <NA>
#> NP_001001801.1_S30s              <NA>          <NA>       <NA>
#> NP_001002016.2_S12s       P02545_S12s        P02545        S12
#> NP_001002016.2_S143s     P02545_S143s        P02545       S143
#> NP_001002016.2_S153s     P02545_S153s        P02545       S153
head(pData(PHOSPHO_MSNSET))      # phenoData
#>                  pid   bid       labelid   viallabel    sex timepoint exp_group
#> 90222017006 10025626 90222 4.457560e-313 90222017006   Male        8W      M_8W
#> 90258017006 10045228 90258 4.459338e-313 90258017006 Female        8W      F_8W
#> 90441017006 10502300 90441 4.468380e-313 90441017006   Male        2W      M_2W
#> 90266017006 10046380 90266 4.459734e-313 90266017006 Female       SED     F_SED
#> 90421017006 10417702 90421 4.467392e-313 90421017006   Male        1W      M_1W
#> 90410017006 10335064 90410 4.466848e-313 90410017006 Female        4W      F_4W