Proteomics MSnSet

An object of class MSnbase::MSnSet containing median-MAD-normalized log\(_2\)-transformed protein expression data for differential analysis and WGCNA.

Usage

PROT_MSNSET

Format

Object of class MSnbase::MSnSet with 9964 features and 60 samples.

assayData

Object of class matrix with 9964 rows and 60 columns. See PROT_WATSC_NORM_DATA for more details.

featureData

Object of class data.frame with 9964 rows and 3 variables containing additional feature information:

feature_ID

character; Reference Sequence (RefSeq) protein identifier.

gene_symbol

character; official gene symbol.

entrez_gene

integer; Entrez gene identifier.

phenoData

Object of class data.frame with 60 rows and 7 variables containing additional sample information:

pid

integer; randomly generated 8-digit identifier used in linkage to phenotypic data. All samples from the same animal have the same PID.

bid

integer; unique 5 digit identifier of all samples collected for an acute test/sample collection period. All samples collected during that period will have the same BID.

labelid

integer; unique 11 digit specimen label identifier, originating at the collection site, that provides a link to specimen processing and is used for shipments to the biorepository. Same as viallabel only in instances where aliquots are not further processed at the biorepository.

viallabel

character; unique 11 digit sample vial identifier. Starts with the bid.

sex

factor; the sex of the rat with levels "Female" and "Male".

timepoint

factor; exercise training group. Either "SED" (sedentary) or the number of weeks of training ("1W", "2W", "4W", "8W").

exp_group

factor; experimental group. Unique combination of sex (first letter) and timepoint.

Details

FEATURE_TO_GENE for details of feature-to-gene mapping.

See also

Examples

library(MSnbase)
PROT_MSNSET                   # display summary
#> MSnSet (storageMode: lockedEnvironment)
#> assayData: 9964 features, 60 samples 
#>   element names: exprs 
#> protocolData: none
#> phenoData
#>   sampleNames: 90222017006 90258017006 ... 90237017006 (60 total)
#>   varLabels: pid bid ... exp_group (7 total)
#>   varMetadata: labelDescription
#> featureData
#>   featureNames: AP_004893.1 AP_004894.1 ... YP_665639.1 (9964 total)
#>   fvarLabels: feature_ID gene_symbol entrez_gene
#>   fvarMetadata: labelDescription
#> experimentData: use 'experimentData(object)'
#> Annotation:  
#> - - - Processing information - - -
#>  MSnbase version: 2.22.0 
exprs(PROT_MSNSET)[1:10, 1:5] # assayData (first 10 rows and 5 columns)
#>                90222017006 90258017006 90441017006 90266017006 90421017006
#> AP_004893.1        0.06250    -0.00673     0.06530    -0.12847     0.42877
#> AP_004894.1       -0.35726    -0.00678    -0.00001    -0.09093     0.34666
#> AP_004895.1       -0.29374     0.16927     0.04853    -0.08151     0.34469
#> AP_004896.1        0.03139    -0.05546     0.06703    -0.09930     0.22900
#> AP_004898.1       -0.20785     0.18767     0.10063    -0.33669     0.15039
#> AP_004899.1             NA          NA          NA          NA          NA
#> AP_004901.1             NA          NA          NA          NA          NA
#> AP_004904.1       -0.12326    -0.35704    -0.24179    -0.01929     0.79072
#> NP_001000516.1    -0.42419     0.11770     0.41690    -0.09733    -0.35746
#> NP_001001505.1    -0.17732    -0.13575    -0.03853     0.15163    -0.19405
head(fData(PROT_MSNSET))      # featureData
#>              feature_ID gene_symbol entrez_gene
#> AP_004893.1 AP_004893.1      Mt-nd2       26194
#> AP_004894.1 AP_004894.1      Mt-co1       26195
#> AP_004895.1 AP_004895.1      Mt-co2       26198
#> AP_004896.1 AP_004896.1     Mt-atp8       26196
#> AP_004898.1 AP_004898.1      Mt-co3       26204
#> AP_004899.1 AP_004899.1      Mt-nd3       26199
head(pData(PROT_MSNSET))      # phenoData
#>                  pid   bid       labelid   viallabel    sex timepoint exp_group
#> 90222017006 10025626 90222 4.457560e-313 90222017006   Male        8W      M_8W
#> 90258017006 10045228 90258 4.459338e-313 90258017006 Female        8W      F_8W
#> 90441017006 10502300 90441 4.468380e-313 90441017006   Male        2W      M_2W
#> 90266017006 10046380 90266 4.459734e-313 90266017006 Female       SED     F_SED
#> 90421017006 10417702 90421 4.467392e-313 90421017006   Male        1W      M_1W
#> 90410017006 10335064 90410 4.466848e-313 90410017006 Female        4W      F_4W