Transcriptomics MSnSet

An MSnSet-class object containing transcriptomics data from rat WAT provided by MoTrPAC.

Usage

TRNSCRPT_MSNSET

Format

An MSnSet-class object with 16443 features and 48 samples.

assayData

A matrix of counts with 16443 rows and 48 columns.

featureData

A data.frame with 30304 rows and 7 variables containing additional feature information:

feature_ID

character;

gene_symbol

character list; the gene symbol associated with each protein.

entrez_gene

numeric list; the Entrez gene identifier associated with each protein.

site

character;

human_feature_ID

character;

human_uniprot

character;

human_site

character;

Prior to collapsing the entrez_id and gene_symbol columns into lists, any rows where gene_symbol began with "LOC" were removed unless that would remove entire features. This reduces the impact of the one-to-many feature-to-gene mapping on FGSEA results.

phenoData

A data.frame with 48 rows and 9 variables containing sample information:

viallabel

character; unique sample identifier.

rin

numeric; RNA integrity number (RIN).

pct_globin

numeric; percent of reads mapping to globin.

pct_umi_dup

numeric; percent of PCR duplicates as quantified with Unique Molecular Identifiers (UMIs).

median_5_3_bias

numeric; median 5'-3' bias.

bid

character; Unique 5 digit numeric identifier of all samples collected for an acute test/sample collection period. All samples collected during that period will have the same BID. Same as first 5 digits of viallabel.

sex

factor; the sex of the rat with 2 levels "Female" and "Male".

timepoint

factor; exercise training group. Either "SED" (sedentary) or the number of weeks of training ("1W", "2W", "4W", "8W").

exp_group

factor; unique combination of sex (first letter) and timepoint.

Details

Two outlier samples ("90423", "90410") were removed (48 remain). Then, pct_globin, rin, pct_umi_dup, and median_5_3_bias were mean-imputed, centered, and scaled.

Examples

TRNSCRPT_MSNSET # summary
#> Loading required package: MSnbase
#> Warning: there is no package called ‘MSnbase’
#> Error in .requirePackage(package): unable to find required package ‘MSnbase’
library(MSnbase)
#> Error in library(MSnbase): there is no package called ‘MSnbase’
exprs(TRNSCRPT_MSNSET)[1:5, 1:5] # assayData
#> Error in exprs(TRNSCRPT_MSNSET): could not find function "exprs"
head(fData(TRNSCRPT_MSNSET))     # featureData
#> Error in fData(TRNSCRPT_MSNSET): could not find function "fData"
head(pData(TRNSCRPT_MSNSET))     # phenoData
#> Error in pData(TRNSCRPT_MSNSET): could not find function "pData"